Convert deploy command to oclif
Change-type: patch
